controller/requirements.txt
uvicorn~=0.38.0
fastapi~=0.124.4
asyncpg~=0.31.0
dotenv~=0.9.9
python-dotenv~=1.2.1
urllib3~=2.6.2
minio~=7.2.20
h11~=0.16.0
pip~=25.0.1
protobuf~=4.21.12
cryptography~=46.0.2
Jinja2~=3.1.6
cffi~=2.0.0
PyYAML~=6.0.3
typing_extensions~=4.15.0
certifi~=2025.11.12
pycryptodome~=3.23.0
argon2-cffi-bindings~=25.1.0
click~=8.3.1
pydantic~=2.12.5
pydantic_core~=2.41.5
starlette~=0.50.0
anyio~=4.12.0
annotated-types~=0.7.0
pycparser~=2.23
itsdangerous~=2.2.0